remove one blockdataform
